requirements.txt
torch
torchvision
torchaudio
open_clip_torch
timm
numpy
pillow
tqdm
matplotlib
